Create a permanently failed output stream from the spout. (#90)
A tuple is considered "permanently failed" when the topology has attempted to process the tuple at least once and the RetryManager implementation has determined that the tuple should not be retried. When this occurs, the tuple will be emitted un-anchored out a "failed" stream. Bolts within the topology can subscribe to this "failed" stream and do its own error handling. The name of this stream is configured via the `spout.permanently_failed_output_stream_id` setting, and if undefined defaults simply to `failed`